Use the process without any children instead of the users shell
to determine the cwd of the window.  This will still work correctly
when there are multiple shells open, and the later shells use a
different cwd than the initial login shell.  Also works when e.g.
the login shell is bash, but the user has invoked zsh from there
manually, and then switched the directory.